requirements.txt
fastapi
uvicorn
python-dotenv
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.0
langchain-core==0.3.15
chromadb
requests
pydantic
numpy==1.26.4